+++ Audentia-Core/src/main/java/fr/audentia/core/infrastructure/repositories/balance/MariaDbBalanceRepository.java
@@ -3,13 +3,13 @@ package fr.audentia.core.infrastructure.repositories.balance;
 import fr.audentia.core.domain.balance.BalanceRepository;
 import fr.audentia.players.domain.teams.Team;
 
-import java.util.OptionalDouble;
+import java.util.OptionalInt;
 
 public class MariaDbBalanceRepository implements BalanceRepository {
 
     @Override
-    public OptionalDouble getTeamBalance(Team team) {
-        return OptionalDouble.empty();
+    public OptionalInt getTeamBalance(Team team) {
+        return OptionalInt.empty();
     }
 
 }
